# ec-service/src/main/java/ru/otus/controller/compilation/AdminCompilationController.java
package ru.otus.controller.compilation;

import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import ru.otus.model.UpdateCompilationRequest;
import ru.otus.dto.CompilationDto;
import ru.otus.dto.NewCompilationDto;
import ru.otus.service.compilation.CompilationService;

import javax.validation.Valid;
import javax.validation.constraints.NotNull;

@Slf4j
@Validated
@RestController
@RequiredArgsConstructor
@RequestMapping("/admin/compilations")
public class AdminCompilationController {

    private final CompilationService compilationService;

    @PostMapping
    public ResponseEntity<CompilationDto> create(@Valid @RequestBody NewCompilationDto newCompilationDto) {
        log.info("Post /admin/compilations with newCompilationDto={}", newCompilationDto);
        return new ResponseEntity<>(compilationService.create(newCompilationDto), HttpStatus.CREATED);
    }

    @DeleteMapping("/{compId}")
    public ResponseEntity<Void> delete(@NotNull @PathVariable Long compId) {
        log.info("Delete /admin/compilations/{}", compId);
        compilationService.delete(compId);
        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
    }

    @PatchMapping("/{compId}")
    public CompilationDto update(@PathVariable Long compId, @Valid @RequestBody UpdateCompilationRequest request) {
        log.info("Patch /admin/compilations/{} with request={}", compId, request);
        return compilationService.update(compId, request);
    }
}