# swagger/ec-main-service-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Main service API",
    "version": "1.0"
  },
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost:8080"
    }
  ],
...
  "paths": {
    "/admin/categories": {
      "post": {
        "description": "Обратите внимание: имя категории должно быть уникальным",
        "operationId": "addCategory",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewCategoryDto"
              }
            }
          },
          "description": "данные добавляемой категории",
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            },
            "description": "Категория добавлена"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "status": "BAD_REQUEST",
                  "reason": "Incorrectly made request.",
                  "message": "Field: name. Error: must not be blank. Value: null",
                  "timestamp": "2022-09-07 09:10:50"
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Запрос составлен некорректно"
          },
          "409": {
            "content": {
              "application/json": {
                "example": {
                  "status": "CONFLICT",
                  "reason": "Integrity constraint has been violated.",
                  "message": "could not execute statement; SQL [n/a]; constraint [uq_category_name]; nested exception is org.hibernate.exception.ConstraintViolationException: could not execute statement",
                  "timestamp": "2022-09-07 09:10:50"
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Нарушение целостности данных"
          }
        },
        "summary": "Добавление новой категории",
        "tags": [
          "Admin: Категории"
        ]
      }
    },
    "/admin/categories/{catId}": {
      "delete": {
        "description": "Обратите внимание: с категорией не должно быть связано ни одного события.",
        "operationId": "deleteCategory",
        "parameters": [
          {
            "description": "id категории",
            "in": "path",
            "name": "catId",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Категория удалена"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "status": "NOT_FOUND",
                  "reason": "The required object was not found.",
                  "message": "Category with id=27 was not found",
                  "timestamp": "2022-09-07 09:10:50"
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Категория не найдена или недоступна"
          },
          "409": {
            "content": {
              "application/json": {
                "example": {
                  "status": "CONFLICT",
                  "reason": "For the requested operation the conditions are not met.",
                  "message": "The category is not empty",
                  "timestamp": "2023-01-21 16:56:19"
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Существуют события, связанные с категорией"
          }
        },
        "summary": "Удаление категории",
        "tags": [
          "Admin: Категории"
        ]
      },
      "patch": {
        "description": "Обратите внимание: имя категории должно быть уникальным",
        "operationId": "updateCategory",
        "parameters": [
          {
            "description": "id категории",
            "in": "path",
            "name": "catId",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryDto"
              }
            }
          },
          "description": "Данные категории для изменения",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            },
            "description": "Данные категории изменены"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "status": "NOT_FOUND",
                  "reason": "The required object was not found.",
                  "message": "Category with id=27 was not found",
                  "timestamp": "2022-09-07 09:10:50"
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Категория не найдена или недоступна"
          },
          "409": {
            "content": {
              "application/json": {
                "example": {
                  "status": "CONFLICT",
                  "reason": "Integrity constraint has been violated.",
                  "message": "could not execute statement; SQL [n/a]; constraint [uq_category_name]; nested exception is org.hibernate.exception.ConstraintViolationException: could not execute statement",
                  "timestamp": "2022-09-07 09:10:50"
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Нарушение целостности данных"
          }
        },
        "summary": "Изменение категории",
        "tags": [
          "Admin: Категории"
        ]
      }
    },
    "/admin/compilations": {
      "post": {
        "operationId": "saveCompilation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewCompilationDto"
              }
            }
          },
          "description": "данные новой подборки",
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompilationDto"
                }
              }
            },
            "description": "Подборка добавлена"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "status": "BAD_REQUEST",
                  "reason": "Incorrectly made request.",
                  "message": "Field: title. Error: must not be blank. Value: null",
                  "timestamp": "2022-09-07 09:10:50"
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Запрос составлен некорректно"
          },
          "409": {
            "content": {
              "application/json": {
                "example": {
                  "status": "CONFLICT",
                  "reason": "Integrity constraint has been violated.",
                  "message": "could not execute statement; SQL [n/a]; constraint [uq_compilation_name]; nested exception is org.hibernate.exception.ConstraintViolationException: could not execute statement",
                  "timestamp": "2022-09-07 09:10:50"
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Нарушение целостности данных"
          }
        },
        "summary": "Добавление новой подборки (подборка может не содержать событий)",
        "tags": [
          "Admin: Подборки событий"
        ]
      }
    },
    "/admin/compilations/{compId}": {
      "delete": {
        "operationId": "deleteCompilation",
        "parameters": [
          {
            "description": "id подборки",
            "in": "path",
            "name": "compId",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Подборка удалена"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "status": "NOT_FOUND",
                  "reason": "The required object was not found.",
                  "message": "Compilation with id=11 was not found",
                  "timestamp": "2022-09-07 09:10:50"
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Подборка не найдена или недоступна"
          }
        },
        "summary": "Удаление подборки",
        "tags": [
          "Admin: Подборки событий"
        ]
      },
      "patch": {
        "operationId": "updateCompilation",
        "parameters": [
          {
            "description": "id подборки",
            "in": "path",
            "name": "compId",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCompilationRequest"
              }
            }
          },
          "description": "данные для обновления подборки",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompilationDto"
                }
              }
            },
            "description": "Подборка обновлена"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "status": "NOT_FOUND",
                  "reason": "The required object was not found.",
                  "message": "Category with id=27 was not found",
                  "timestamp": "2022-09-07 09:10:50"
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Подборка не найдена или недоступна"
          }
        },
        "summary": "Обновить информацию о подборке",
        "tags": [
          "Admin: Подборки событий"
        ]
      }
    },
...
    "/categories": {
      "get": {
        "description": "В случае, если по заданным фильтрам не найдено ни одной категории, возвращает пустой список",
        "operationId": "getCategories",
        "parameters": [
          {
            "description": "количество категорий, которые нужно пропустить для формирования текущего набора",
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "description": "количество категорий в наборе",
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryDto"
                  }
                }
              }
            },
            "description": "Категории найдены"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "status": "BAD_REQUEST",
                  "reason": "Incorrectly made request.",
                  "message": "Failed to convert value of type java.lang.String to required type int; nested exception is java.lang.NumberFormatException: For input string: ad",
                  "timestamp": "2022-09-07 09:10:50"
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Запрос составлен некорректно"
          }
        },
        "summary": "Получение категорий",
        "tags": [
          "Public: Категории"
        ]
      }
    },
    "/categories/{catId}": {
      "get": {
        "description": "В случае, если категории с заданным id не найдено, возвращает статус код 404",
        "operationId": "getCategory",
        "parameters": [
          {
            "description": "id категории",
            "in": "path",
            "name": "catId",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            },
            "description": "Категория найдена"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "status": "BAD_REQUEST",
                  "reason": "Incorrectly made request.",
                  "message": "Failed to convert value of type java.lang.String to required type long; nested exception is java.lang.NumberFormatException: For input string: ad",
                  "timestamp": "2022-09-07 09:10:50"
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Запрос составлен некорректно"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "status": "NOT_FOUND",
                  "reason": "The required object was not found.",
                  "message": "Category with id=19 was not found",
                  "timestamp": "2022-09-07 09:10:50"
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Категория не найдена или недоступна"
          }
        },
        "summary": "Получение информации о категории по её идентификатору",
        "tags": [
          "Public: Категории"
        ]
      }
    },
...
    "/compilations/{compId}": {
      "get": {
        "description": "В случае, если подборки с заданным id не найдено, возвращает статус код 404",
        "operationId": "getCompilation",
        "parameters": [
          {
            "description": "id подборки",
            "in": "path",
            "name": "compId",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompilationDto"
                }
              }
            },
            "description": "Подборка событий найдена"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "status": "BAD_REQUEST",
                  "reason": "Incorrectly made request.",
                  "message": "Failed to convert value of type java.lang.String to required type long; nested exception is java.lang.NumberFormatException: For input string: ad",
                  "timestamp": "2022-09-07 09:10:50"
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Запрос составлен некорректно"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "status": "NOT_FOUND",
                  "reason": "The required object was not found.",
                  "message": "Compilation with id=84 was not found",
                  "timestamp": "2022-09-07 09:10:50"
                },
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Подборка не найдена или недоступна"
          }
        },
        "summary": "Получение подборки событий по его id",
        "tags": [
          "Public: Подборки событий"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "array",
            "description": "Список стектрейсов или описания ошибок",
            "example": [],
            "items": {
              "type": "string",
              "description": "Список стектрейсов или описания ошибок",
              "example": "[]"
            }
          },
          "message": {
            "type": "string",
            "description": "Сообщение об ошибке",
            "example": "Only pending or canceled events can be changed"
          },
          "reason": {
            "type": "string",
            "description": "Общее описание причины ошибки",
            "example": "For the requested operation the conditions are not met."
          },
          "status": {
            "type": "string",
            "description": "Код статуса HTTP-ответа",
            "example": "FORBIDDEN",
            "enum": [
              "100 CONTINUE",
              "101 SWITCHING_PROTOCOLS",
              "102 PROCESSING",
              "103 CHECKPOINT",
              "200 OK",
              "201 CREATED",
              "202 ACCEPTED",
              "203 NON_AUTHORITATIVE_INFORMATION",
              "204 NO_CONTENT",
              "205 RESET_CONTENT",
              "206 PARTIAL_CONTENT",
              "207 MULTI_STATUS",
              "208 ALREADY_REPORTED",
              "226 IM_USED",
              "300 MULTIPLE_CHOICES",
              "301 MOVED_PERMANENTLY",
              "302 FOUND",
              "302 MOVED_TEMPORARILY",
              "303 SEE_OTHER",
              "304 NOT_MODIFIED",
              "305 USE_PROXY",
              "307 TEMPORARY_REDIRECT",
              "308 PERMANENT_REDIRECT",
              "400 BAD_REQUEST",
              "401 UNAUTHORIZED",
              "402 PAYMENT_REQUIRED",
              "403 FORBIDDEN",
              "404 NOT_FOUND",
              "405 METHOD_NOT_ALLOWED",
              "406 NOT_ACCEPTABLE",
              "407 PROXY_AUTHENTICATION_REQUIRED",
              "408 REQUEST_TIMEOUT",
              "409 CONFLICT",
              "410 GONE",
              "411 LENGTH_REQUIRED",
              "412 PRECONDITION_FAILED",
              "413 PAYLOAD_TOO_LARGE",
              "413 REQUEST_ENTITY_TOO_LARGE",
              "414 URI_TOO_LONG",
              "414 REQUEST_URI_TOO_LONG",
              "415 UNSUPPORTED_MEDIA_TYPE",
              "416 REQUESTED_RANGE_NOT_SATISFIABLE",
              "417 EXPECTATION_FAILED",
              "418 I_AM_A_TEAPOT",
              "419 INSUFFICIENT_SPACE_ON_RESOURCE",
              "420 METHOD_FAILURE",
              "421 DESTINATION_LOCKED",
              "422 UNPROCESSABLE_ENTITY",
              "423 LOCKED",
              "424 FAILED_DEPENDENCY",
              "425 TOO_EARLY",
              "426 UPGRADE_REQUIRED",
              "428 PRECONDITION_REQUIRED",
              "429 TOO_MANY_REQUESTS",
              "431 REQUEST_HEADER_FIELDS_TOO_LARGE",
              "451 UNAVAILABLE_FOR_LEGAL_REASONS",
              "500 INTERNAL_SERVER_ERROR",
              "501 NOT_IMPLEMENTED",
              "502 BAD_GATEWAY",
              "503 SERVICE_UNAVAILABLE",
              "504 GATEWAY_TIMEOUT",
              "505 HTTP_VERSION_NOT_SUPPORTED",
              "506 VARIANT_ALSO_NEGOTIATES",
              "507 INSUFFICIENT_STORAGE",
              "508 LOOP_DETECTED",
              "509 BANDWIDTH_LIMIT_EXCEEDED",
              "510 NOT_EXTENDED",
              "511 NETWORK_AUTHENTICATION_REQUIRED"
            ]
          },
          "timestamp": {
            "type": "string",
            "description": "Дата и время когда произошла ошибка (в формате \"yyyy-MM-dd HH:mm:ss\")",
            "example": "2022-06-09 06:27:23"
          }
        },
        "description": "Сведения об ошибке"
      },
      "CategoryDto": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Идентификатор категории",
            "format": "int64",
            "readOnly": true,
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Название категории",
            "example": "Концерты"
          }
        },
        "description": "Категория"
      },
      "CompilationDto": {
        "required": [
          "id",
          "pinned",
          "title"
        ],
        "type": "object",
        "properties": {
          "events": {
            "uniqueItems": true,
            "type": "array",
            "description": "Список событий входящих в подборку",
            "example": [
              {
                "annotation": "Эксклюзивность нашего шоу гарантирует привлечение максимальной зрительской аудитории",
                "category": {
                  "id": 1,
                  "name": "Концерты"
                },
                "confirmedRequests": 5,
                "eventDate": "2024-03-10 14:30:00",
                "id": 1,
                "initiator": {
                  "id": 3,
                  "name": "Фёдоров Матвей"
                },
                "paid": true,
                "title": "Знаменитое шоу 'Летающая кукуруза'",
                "views": 999
              },
              {
                "annotation": "За почти три десятилетия группа 'Java Core' закрепились на сцене как группа, объединяющая поколения.",
                "category": {
                  "id": 1,
                  "name": "Концерты"
                },
                "confirmedRequests": 555,
                "eventDate": "2025-09-13 21:00:00",
                "id": 1,
                "initiator": {
                  "id": 3,
                  "name": "Паша Петров"
                },
                "paid": true,
                "title": "Концерт рок-группы 'Java Core'",
                "views": 991
              }
            ],
            "items": {
              "$ref": "#/components/schemas/EventShortDto"
            }
          },
          "id": {
            "type": "integer",
            "description": "Идентификатор",
            "format": "int64",
            "example": 1
          },
          "pinned": {
            "type": "boolean",
            "description": "Закреплена ли подборка на главной странице сайта",
            "example": true
          },
          "title": {
            "type": "string",
            "description": "Заголовок подборки",
            "example": "Летние концерты"
          }
        },
        "description": "Подборка событий"
      },
...
      "EventShortDto": {
        "required": [
          "annotation",
          "category",
          "eventDate",
          "initiator",
          "paid",
          "title"
        ],
        "type": "object",
        "properties": {
          "annotation": {
            "type": "string",
            "description": "Краткое описание",
            "example": "Эксклюзивность нашего шоу гарантирует привлечение максимальной зрительской аудитории"
          },
          "category": {
            "$ref": "#/components/schemas/CategoryDto"
          },
          "confirmedRequests": {
            "type": "integer",
            "description": "Количество одобренных заявок на участие в данном событии",
            "format": "int64",
            "example": 5
          },
          "eventDate": {
            "type": "string",
            "description": "Дата и время на которые намечено событие (в формате \"yyyy-MM-dd HH:mm:ss\")",
            "example": "2024-12-31 15:10:05"
          },
          "id": {
            "type": "integer",
            "description": "Идентификатор",
            "format": "int64",
            "example": 1
          },
          "initiator": {
            "$ref": "#/components/schemas/UserShortDto"
          },
          "paid": {
            "type": "boolean",
            "description": "Нужно ли оплачивать участие",
            "example": true
          },
          "title": {
            "type": "string",
            "description": "Заголовок",
            "example": "Знаменитое шоу 'Летающая кукуруза'"
          },
          "views": {
            "type": "integer",
            "description": "Количество просмотрев события",
            "format": "int64",
            "example": 999
          }
        },
        "description": "Краткая информация о событии",
        "example": [
          {
            "annotation": "Эксклюзивность нашего шоу гарантирует привлечение максимальной зрительской аудитории",
            "category": {
              "id": 1,
              "name": "Концерты"
            },
            "confirmedRequests": 5,
            "eventDate": "2024-03-10 14:30:00",
            "id": 1,
            "initiator": {
              "id": 3,
              "name": "Фёдоров Матвей"
            },
            "paid": true,
            "title": "Знаменитое шоу 'Летающая кукуруза'",
            "views": 999
          },
          {
            "annotation": "За почти три десятилетия группа 'Java Core' закрепились на сцене как группа, объединяющая поколения.",
            "category": {
              "id": 1,
              "name": "Концерты"
            },
            "confirmedRequests": 555,
            "eventDate": "2025-09-13 21:00:00",
            "id": 1,
            "initiator": {
              "id": 3,
              "name": "Паша Петров"
            },
            "paid": true,
            "title": "Концерт рок-группы 'Java Core'",
            "views": 991
          }
        ]
      },
...
      "NewCategoryDto": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Название категории",
            "example": "Концерты"
          }
        },
        "description": "Данные для добавления новой категории"
      },
      "NewCompilationDto": {
        "required": [
          "title"
        ],
        "type": "object",
        "properties": {
          "events": {
            "uniqueItems": true,
            "type": "array",
            "description": "Список идентификаторов событий входящих в подборку",
            "example": [
              1,
              2,
              3
            ],
            "items": {
              "type": "integer",
              "description": "Список идентификаторов событий входящих в подборку",
              "format": "int64"
            }
          },
          "pinned": {
            "type": "boolean",
            "description": "Закреплена ли подборка на главной странице сайта",
            "example": false,
            "default": false
          },
          "title": {
            "type": "string",
            "description": "Заголовок подборки",
            "example": "Летние концерты"
          }
        },
        "description": "Подборка событий"
      },
...
      "UpdateCompilationRequest": {
        "type": "object",
        "properties": {
          "events": {
            "uniqueItems": true,
            "type": "array",
            "description": "Список id событий подборки для полной замены текущего списка",
            "items": {
              "type": "integer",
              "description": "Список id событий подборки для полной замены текущего списка",
              "format": "int64"
            }
          },
          "pinned": {
            "type": "boolean",
            "description": "Закреплена ли подборка на главной странице сайта",
            "example": true
          },
          "title": {
            "type": "string",
            "description": "Заголовок подборки",
            "example": "Необычные фотозоны"
          }
        },
        "description": "Изменение информации о подборке событий. Если поле в запросе не указано (равно null) - значит изменение этих данных не треубется."
      },
...
      "UserShortDto": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Идентификатор",
            "format": "int64",
            "example": 3
          },
          "name": {
            "type": "string",
            "description": "Имя",
            "example": "Фёдоров Матвей"
          }
        },
        "description": "Пользователь (краткая информация)"
      }
    }
  }
}